requirements.txt
APScheduler==3.6.3
boto3==1.16.10
botocore==1.19.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
DateTime==4.3
dnspython==1.16.0
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-Cors==3.0.9
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.4
psycopg2-binary==2.8.6
pymongo==3.11.1
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
tweepy==3.9.0
tzlocal==2.1
urllib3==1.25.11
websocket-client==0.57.0
Werkzeug==1.0.1
zope.interface==5.1.2